# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/categorias": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Obtém todas as categorias",
                "description": "Retorna uma lista com todas as categorias cadastradas",
                "operationId": "getCategorias",
                "responses": {
                    "200": {
                        "description": "Lista de categorias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nome": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Cria uma nova categoria",
                "description": "Cria uma nova categoria com o nome fornecido",
                "operationId": "createCategoria",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nome": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados fornecidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista todos os clientes",
                "description": "Retorna uma lista de todos os clientes cadastrados.",
                "operationId": "getClientes",
                "responses": {
                    "200": {
                        "description": "Lista de clientes retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cliente"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a solicitação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cria um novo cliente",
                "description": "Cria um novo cliente com os dados fornecidos",
                "operationId": "createCliente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "idade",
                                    "email",
                                    "cpf"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "idade": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "cpf": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nome": {
                                            "type": "string"
                                        },
                                        "idade": {
                                            "type": "integer"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "cpf": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados fornecidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pagamentos": {
            "get": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Obtém todos os pagamentos",
                "description": "Retorna uma lista com todos os pagamentos cadastrados",
                "operationId": "getPagamentos",
                "responses": {
                    "200": {
                        "description": "Lista de pagamentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nome": {
                                                "type": "string"
                                            },
                                            "id_pedido": {
                                                "type": "integer"
                                            },
                                            "id_tipopagamento": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Cria um novo pagamento",
                "description": "Cria um novo pagamento com os dados fornecidos",
                "operationId": "createPagamento",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_pedido",
                                    "id_tipopagamento"
                                ],
                                "properties": {
                                    "id_pedido": {
                                        "type": "integer"
                                    },
                                    "id_tipopagamento": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pagamento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "id_pedido": {
                                            "type": "integer"
                                        },
                                        "id_tipopagamento": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados fornecidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cliente": {
                "title": "Cliente",
                "description": "Modelo de Cliente",
                "required": [
                    "nome",
                    "email",
                    "cpf"
                ],
                "properties": {
                    "nome": {
                        "description": "Nome do cliente",
                        "type": "string"
                    },
                    "idade": {
                        "description": "Idade do cliente",
                        "type": "integer"
                    },
                    "email": {
                        "description": "Email do cliente",
                        "type": "string",
                        "format": "email"
                    },
                    "cpf": {
                        "description": "CPF do cliente (apenas números)",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}